requirements.txt
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0
pyaudio>=0.2.11
numpy>=1.21.0
tkinter  # Usually comes with Python
scipy>=1.7.0